+++ src/main/java/com/lostway/eventmanager/repository/UserEventRegistrationEntityRepository.java
@@ -7,10 +7,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface UserEventRegistrationEntityRepository extends JpaRepository<UserEventRegistrationEntity, Integer> {
-    boolean existsByUserIdAndEventId(Long userId, Integer eventId);
+public interface UserEventRegistrationEntityRepository extends JpaRepository<UserEventRegistrationEntity, Long> {
+    boolean existsByUserIdAndEventId(Long userId, Long eventId);
 
-    Optional<UserEventRegistrationEntity> findByUserIdAndEventId(Long userId, Integer eventId);
+    Optional<UserEventRegistrationEntity> findByUserIdAndEventId(Long userId, Long eventId);
 
     List<UserEventRegistrationEntity> findByUser(UserEntity user);
 }
